requirements.txt
# Well Mind Web Application - Python Dependencies

# Core Flask framework
Flask==2.3.3
Flask-CORS==4.0.0

# Database
mysql-connector-python==8.1.0

# Security and Authentication
Werkzeug==2.3.7
PyJWT==2.8.0

# AI and Natural Language Processing
openai==0.28.1
textblob==0.17.1

# Data Processing
pandas==2.1.1
numpy==1.24.3

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# JSON handling
jsonschema==4.19.1

# Logging
colorlog==6.7.0

# Development and Testing
pytest==7.4.2
pytest-flask==1.2.0

# Additional utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# For production deployment (optional)
gunicorn==21.2.0